# Conflicts:
#	src/main/java/com/example/scheduleappdev/controller/ScheduleController.java
#	src/main/java/com/example/scheduleappdev/controller/UserController.java
#	src/main/java/com/example/scheduleappdev/dto/CreateScheduleReqDto.java
#	src/main/java/com/example/scheduleappdev/dto/CreateUserReqDto.java
#	src/main/java/com/example/scheduleappdev/dto/LoginUserReqDto.java
#	src/main/java/com/example/scheduleappdev/dto/UpdateScheduleReqDto.java
#	src/main/java/com/example/scheduleappdev/dto/UpdateUserPasswordReqDto.java
#	src/main/java/com/example/scheduleappdev/exception/GlobalExceptionHandler.java
#	src/main/java/com/example/scheduleappdev/repository/UserRepository.java
#	src/main/java/com/example/scheduleappdev/service/ScheduleService.java
#	src/main/java/com/example/scheduleappdev/service/UserService.java

+++ src/main/java/com/example/scheduleappdev/service/UserService.java
@@ -1,10 +1,13 @@
 package com.example.scheduleappdev.service;
 
 import com.example.scheduleappdev.config.PasswordEncoder;
+import com.example.scheduleappdev.common.Const;
 import com.example.scheduleappdev.dto.UserResDto;
 import com.example.scheduleappdev.entity.User;
 import com.example.scheduleappdev.exception.UnauthorizedException;
 import com.example.scheduleappdev.repository.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -21,6 +24,7 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
 
     public UserResDto createUser(String userName, String userEmail, String password) {
+        userRepository.isExistUserNameOrEmail(userName, userEmail);
         String encodedPassword = passwordEncoder.encode(password);
         User user = new User(userName, userEmail, encodedPassword);
         log.info("유저 생성 : name = {}", userName);
@@ -34,6 +38,20 @@ public class UserService {
             throw new UnauthorizedException("비밀번호가 일치하지 않습니다.");
         }
         return new UserResDto(user);
+    }
+
+    public void makeSession(HttpServletRequest req, UserResDto userResDto) {
+        HttpSession session = req.getSession();
+        session.setAttribute(Const.LOGIN_USER, userResDto);
+        log.info("로그인 성공 : name = {}", userResDto.getUserName());
+    }
+
+    public void logout(HttpServletRequest req) {
+        HttpSession session = req.getSession(false);
+        if (session != null) {
+            session.invalidate();
+        }
+        log.info("로그아웃");
     }
 
     public List<UserResDto> findAllUsers() {
@@ -61,10 +79,21 @@ public class UserService {
         return new UserResDto(findUser);
     }
 
-    public void deleteUserById(Long id) {
-        User findUser = userRepository.findByIdOrElseThrow(id);
-        log.info("유저 삭제 : name = {}", findUser.getUserName());
-        userRepository.delete(findUser);
+    public void deleteUser(String password, HttpServletRequest req) {
+        HttpSession session = req.getSession(false);
+        if (session != null) {
+            UserResDto loginUser = (UserResDto) session.getAttribute(Const.LOGIN_USER);
+            User findUser = userRepository.findByIdOrElseThrow(loginUser.getUserId());
+
+            if (!findUser.getPassword().equals(password)) {
+                throw new UnauthorizedException("비밀번호가 일치하지 않습니다.");
+            }
+            log.info("회원탈퇴 : name = {}", findUser.getUserName());
+            userRepository.delete(findUser);
+
+            log.info("로그아웃");
+            session.invalidate();
+        }
     }
 
 }
